Nadia review - created folder multidomain in config/datasets and config/prompts with all new datasets - moved and tested dataset processors to separate multidomain processor file - code rag bench: added and tested separate configs per doc dataset - removed local model checkpoint config - removed comments in config_query_generator/unfold_api_query.yaml - added optional support for wandb (and tested) + added training documentation to use / disable wandb - num_saving_steps now config argument
TODO: debug ignored error in vllm -> __del__ -> gc.collect()
