--- config/dataset/apibench_HF.yaml
@@ -1,9 +1,0 @@
-dev:
-    doc:
-        init_args:
-            _target_: modules.dataset_processor.API_gorilla_HF
-            split: eval
-    query: 
-        init_args:
-            _target_: modules.dataset_processor.APIBench_gorilla_HF
-            split: eval
