+++ config/dataset/multidomain/apibench_HF.yaml
@@ -1,0 +1,9 @@
+dev:
+    doc:
+        init_args:
+            _target_: modules.processors.multidomain_dataset_processor.API_gorilla_HF
+            split: eval
+    query: 
+        init_args:
+            _target_: modules.processors.multidomain_dataset_processor.APIBench_gorilla_HF
+            split: eval
