+++ config/dataset/multidomain/apibench_TF.yaml
@@ -1,0 +1,9 @@
+dev:
+    doc:
+        init_args:
+            _target_: modules.processors.multidomain_dataset_processor.API_gorilla_TF
+            split: eval
+    query: 
+        init_args:
+            _target_: modules.processors.multidomain_dataset_processor.APIBench_gorilla_TF
+            split: eval
