+++ config/dataset/multidomain/pubmed_bioasq_ragged.yaml
@@ -1,0 +1,12 @@
+dev:
+    doc: 
+        init_args:
+            _target_: modules.processors.multidomain_dataset_processor.PubMed2023_Ragged
+            split: "train"
+    query: 
+        init_args:
+            _target_: modules.processors.multidomain_dataset_processor.BIOASQ11B_Ragged
+            split: "train"
+test:
+    doc: null
+    query: null
