--- config/dataset/pubmed_bioasq_ragged.yaml
@@ -1,12 +1,0 @@
-dev:
-    doc: 
-        init_args:
-            _target_: modules.dataset_processor.PubMed2023_Ragged
-            split: "train"
-    query: 
-        init_args:
-            _target_: modules.dataset_processor.BIOASQ11B_Ragged
-            split: "train"
-test:
-    doc: null
-    query: null
