+++ config/prompt/multidomain/syllabusqa.yaml
